% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlluminaMouse1Methylationmanifest-package.R
\docType{package}
\name{IlluminaMouse1Methylationmanifest-package}
\alias{IlluminaMouse1Methylationmanifest}
\alias{IlluminaMouse1Methylationmanifest-package}
\title{IlluminaMouse1Methylationmanifest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lihua Huang \email{atomnhuang@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
