% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlluminaMouse1Methylationmanifest.R
\docType{data}
\name{IlluminaMouse1Methylationmanifest}
\alias{IlluminaMouse1Methylationmanifest}
\title{Data of manifest for illumina mouse methylation beadchip on 02/11/2021
create the manifest in reference of IlluminaHumanMethylationEPICmanifest.}
\format{
A DataFrame with 5 sub DataFrames:
\describe{
\item{\verb{type I probes}}{Number of type I probes: 142262 }
\item{\verb{type II probes}}{Number of type II probes: 724574}
\item{\verb{control probes}}{Number of control probes: 635 }
\item{\verb{SNP type I probes}}{Number of SNP type I probes: 21}
\item{\verb{SNP type II probes}}{Number of SNP type II probes: 38}
}
}
\source{
\url{https://support.illumina.com/content/dam/illumina-support/documents/downloads/productfiles/mouse-methylation/infinium-mouse-methylation-manifest-file-csv.zip}
}
\usage{
IlluminaMouse1Methylationmanifest
}
\description{
Data of manifest for illumina mouse methylation beadchip on 02/11/2021
create the manifest in reference of IlluminaHumanMethylationEPICmanifest.
}
\keyword{datasets}
